refactor: Move 'server' to 'infrastructure'
